// frontend-dealers/src/app/(authenticated)/car-evaluation/_components/ExteriorSection/_components/RareRight/types.ts
export type Condition =
  | "Good"
  | "Scratched"
  | "Dented"
  | "Damaged"
  | "Repaired"
  | "Replaced"
  | "Rusted"
  | "Repainted"
  | "Not Available"
  | "Welded"
  | "Spots"
  | "Cracked"
  | "Non-Functional"
  | "Not Working"
  | "Non-Functional"
  | "Clamped"
  | "Chipped"
  | "Faded"
  | "Damaged"
  | "Not Available"
  | "Foggy"
  | "Working"
  | "Color Mismatch"
  | "0-25%"
  | "25-50%"
  | "50-75%"
  | "75-100%"
  | "Missing"
  | "Folding Motor Not Working"
  | "Other";